단위 테스트 후 오류사항 수정 1) 터미널 추가는 구분 변경 가능, 코너추가는 구분 [VAN]으로 고정 2) 코너삭제는 해당 코너에 속한 모든 코너터미널 정보가 제됐을 때만, 삭제되도록 수정 3) 매장터미널 - 터미널 구분 콤보박스 변경 즉시, 저장 및 리스트 재조회 4) 기존 매장중 포스터미널 정보가 없던 매장이 매장정보를 수정하면, 포스 터미널 정보 INSERT 5) 대표밴을 포함하여 터미널 정보 삭제 시, 대표밴만 제외하고 나머지는 삭제처리 되도록 수정

+++ src/main/java/kr/co/solbipos/store/manage/terminalManage/service/impl/TerminalManageServiceImpl.java
@@ -231,7 +231,13 @@ public class TerminalManageServiceImpl implements TerminalManageService{
                 result += mapper.deleteCornerTerminalInfo(storeTerminalVO);
 
                 if("1".equals(storeTerminalVO.getCornrRnum()) && !"Y".equals(storeTerminalVO.getBaseVanYn())) {
-                    result += mapper.deleteCorner(storeCornerVO); // 코너 삭제
+
+                    // 코너 터미널 정보가 남아 있는지 확인하고, 없으면 코너 삭제
+                    List<DefaultMap<String>> list = mapper.getCornerTerminalList(storeTerminalVO);
+
+                    if(list.size() == 0) {
+                        result += mapper.deleteCorner(storeCornerVO); // 코너 삭제
+                    }
                 }
             }
         }
